improve design and bug fixes
