+++ common/engine/src/main/kotlin/com/benjishults/bitnots/engine/proof/FolTableau.kt
@@ -1,0 +1,30 @@
+package com.benjishults.bitnots.engine.proof
+
+import com.benjishults.bitnots.engine.proof.closer.UnifyingClosedIndicator
+import com.benjishults.bitnots.engine.proof.strategy.FolStepStrategy
+import com.benjishults.bitnots.engine.proof.strategy.FolUnificationClosingStrategy
+import com.benjishults.bitnots.engine.proof.strategy.PropositionalInitializationStrategy
+import com.benjishults.bitnots.engine.proof.strategy.StepStrategy
+
+class FolTableau(
+        root: FolTableauNode,
+        stepStrategy: StepStrategy<Tableau> = FolStepStrategy { sf, p: FolTableauNode ->
+            FolTableauNode(sf, p, PropositionalInitializationStrategy())
+        }
+) : AbstractTableau(root, FolUnificationClosingStrategy(), { UnifyingClosedIndicator( it) }, stepStrategy) {
+
+    override fun toString(): String {
+        return buildString {
+            root.preOrderWithPath<PropositionalTableauNode> { n, path ->
+                this.append(path.joinToString("."))
+                this.append("\n")
+                this.append(n.toString())
+//                this.append(n.initialClosers.toString())
+                this.append("\n")
+                this.append("\n")
+                false
+            }
+        }
+    }
+
+}
