+++ common/engine/src/main/kotlin/com/benjishults/bitnots/engine/proof/strategy/StepStrategy.kt
@@ -1,0 +1,7 @@
+package com.benjishults.bitnots.engine.proof.strategy
+
+import com.benjishults.bitnots.engine.proof.Tableau
+
+interface StepStrategy<in T : Tableau> {
+    fun step(tableau: T): Boolean
+}
