+++ common/engine/src/main/kotlin/com/benjishults/bitnots/engine/proof/strategy/initialization_strateties.kt
@@ -10,6 +10,7 @@ interface InitializingStrategy {
     fun init(node: TableauNode)
 }
 
+// TODO refactor
 interface PropositionalInitializingStrategy : InitializingStrategy {
     override fun init(node: TableauNode) {
         with(node) {
