+++ common/engine/src/main/kotlin/com/benjishults/bitnots/engine/proof/tableau.kt
@@ -1,17 +1,16 @@
 package com.benjishults.bitnots.engine.proof
 
-import com.benjishults.bitnots.engine.proof.strategy.BooleanClosedIndicator
+import com.benjishults.bitnots.engine.proof.closer.BooleanClosedIndicator
+import com.benjishults.bitnots.engine.proof.closer.InProgressTableauClosedIndicator
+import com.benjishults.bitnots.engine.proof.closer.NotCompatible
+import com.benjishults.bitnots.engine.proof.closer.UnifyingClosedIndicator
 import com.benjishults.bitnots.engine.proof.strategy.FolStepStrategy
 import com.benjishults.bitnots.engine.proof.strategy.FolUnificationClosingStrategy
-import com.benjishults.bitnots.engine.proof.strategy.InProgressTableauClosedIndicator
-import com.benjishults.bitnots.engine.proof.strategy.NotCompatible
 import com.benjishults.bitnots.engine.proof.strategy.PropositionalClosingStrategy
 import com.benjishults.bitnots.engine.proof.strategy.PropositionalInitializationStrategy
 import com.benjishults.bitnots.engine.proof.strategy.PropositionalStepStrategy
 import com.benjishults.bitnots.engine.proof.strategy.StepStrategy
 import com.benjishults.bitnots.engine.proof.strategy.TableauClosingStrategy
-import com.benjishults.bitnots.engine.proof.strategy.UnifyingClosedIndicator
-import com.benjishults.bitnots.model.unifier.EmptySub
 import java.util.Stack
 
 interface Tableau {
@@ -56,15 +55,16 @@ abstract class AbstractTableau(
         // will this ever exceed size 1?
         val toBeExtended = Stack<InProgressTableauClosedIndicator>().apply { push(closedIndicatorFactory(root)) }
         do {
-            val toExtend = toBeExtended.pop()
-            toExtend.nextNode().let { node ->
-                node.branchClosers.forEach { bc ->
+            toBeExtended.pop().let { toExtend ->
+                toExtend.nextNode().branchClosers.forEach { bc ->
                     toExtend.createExtension(bc).let { ext ->
-                        ext.takeIf {
-                            it.isCloser()
-                        }?.let {
-                            return it
-                        } ?: toBeExtended.push(ext) // this means I'll come back to it
+                        if (ext !== NotCompatible) {
+                            ext.takeIf {
+                                it.isCloser()
+                            }?.let {
+                                return it
+                            } ?: toBeExtended.push(ext) // this means I'll come back to it
+                        }
                     }
                 }
                 toExtend.progress().takeIf {
@@ -79,50 +79,4 @@ abstract class AbstractTableau(
 
     override fun step(): Boolean =
             stepStrategy.step(this)
-}
-
-class PropositionalTableau(
-        root: PropositionalTableauNode,
-        stepStrategy: StepStrategy<Tableau> = PropositionalStepStrategy<PropositionalTableauNode> { sf, p: PropositionalTableauNode ->
-            PropositionalTableauNode(sf, p, PropositionalInitializationStrategy())
-        }
-) : AbstractTableau(root, PropositionalClosingStrategy, { BooleanClosedIndicator(it) }, stepStrategy) {
-
-    override fun toString(): String {
-        return buildString {
-            root.preOrderWithPath<PropositionalTableauNode> { n, path ->
-                this.append(path.joinToString("."))
-                this.append("\n")
-                this.append(n.toString())
-//                this.append(n.initialClosers.toString())
-                this.append("\n")
-                this.append("\n")
-                false
-            }
-        }
-    }
-
-}
-
-class FolTableau(
-        root: FolTableauNode,
-        stepStrategy: StepStrategy<Tableau> = FolStepStrategy { sf, p: FolTableauNode ->
-            FolTableauNode(sf, p, PropositionalInitializationStrategy())
-        }
-) : AbstractTableau(root, FolUnificationClosingStrategy(), { UnifyingClosedIndicator( it) }, stepStrategy) {
-
-    override fun toString(): String {
-        return buildString {
-            root.preOrderWithPath<PropositionalTableauNode> { n, path ->
-                this.append(path.joinToString("."))
-                this.append("\n")
-                this.append(n.toString())
-//                this.append(n.initialClosers.toString())
-                this.append("\n")
-                this.append("\n")
-                false
-            }
-        }
-    }
-
 }
